Placed all draft models in one file.
