+++ models/__init__.py
@@ -1,6 +1,5 @@
 from .artist import Artist
 from .lyrics import Lyrics
-from .song import Song
 from .utils import Utils, TekstowoSession
 from .search import ArtistSearch, SongSearch
 from .user import User
